% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_package}
\alias{install_package}
\title{Install packages from CRAN or Bioconductor}
\usage{
install_package(pkg)
}
\arguments{
\item{pkg}{A length-one character vector containing package name}
}
\description{
A convenient function to install R packages from CRAN or
Bioconductor. It tries both repositories and throws failure if
failed in both cases.
}
