% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_package}
\alias{load_package}
\title{Load a package}
\usage{
load_package(pkg)
}
\arguments{
\item{pkg}{A string for package name}
}
\description{
Load a package, and install it if not installed yet
}
