% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{region_overlap}
\alias{region_overlap}
\title{Region overlap}
\usage{
region_overlap(
  x,
  y,
  regionCols1 = 1:3L,
  regionCols2 = 1:3L,
  type = c("any", "within", "start", "end", "equal"),
  mult = c("all", "first", "last"),
  matchedOnly = F
)
}
\arguments{
\item{x}{The first input genomic regions. A \code{\link[=data.frame]{data.frame()}} or
\code{\link[=data.table]{data.table()}}.}

\item{y}{The second input genomic regions. A \code{\link[=data.frame]{data.frame()}} or
\code{\link[=data.table]{data.table()}}.}

\item{regionCols1}{A vector of column names or numbers in \code{x} giving
'chr', 'start', and 'end' columns in order. Default is the first
3 columns.}

\item{regionCols2}{A vector of column names or numbers in \code{y} giving
'chr', 'start', and 'end' columns in order. Default is the first
3 columns.}

\item{type}{ Default value is \code{any}. Allowed values are \code{any},
\code{within}, \code{start}, \code{end} and \code{equal}.

The types shown here are identical in functionality to the function
\code{findOverlaps} in the bioconductor package \code{IRanges}. Let \code{[a,b]}
and \code{[c,d]} be intervals in \code{x} and \code{y} with \code{a<=b} and
\code{c<=d}. For \code{type="start"}, the intervals overlap iff \code{a == c}.
For \code{type="end"}, the intervals overlap iff \code{b == d}. For
\code{type="within"}, the intervals overlap iff \code{a>=c and b<=d}. For
\code{type="equal"}, the intervals overlap iff \code{a==c and b==d}. For
\code{type="any"}, as long as \code{c<=b and d>=a}, they overlap. In addition
to these requirements, they also have to satisfy the \code{minoverlap} argument
as explained above.

NB: \code{maxgap} argument, when > 0, is to be interpreted according to the type
of the overlap. This will be updated once \code{maxgap} is implemented.}

\item{mult}{ When multiple rows in \code{y} match to the row in \code{x},
\code{mult=.} controls which values are returned - \code{"all"} (default),
\code{"first"} or \code{"last"}.}

\item{matchedOnly}{A logical. If TRUE, then non-overlaped input
regions won't be output. If FALSE (default), all the rows are
output}
}
\value{
A data.table. The columns from \code{x} and \code{y} will have
suffix '.x' and '.y', respectively.
}
\description{
Given two lists of genomic regions, a \linkS4class{data.table} containing
the overlapped regions are returned.
}
\details{
The input genomic regions should contain 3 columns, 'chr', 'start',
and 'end', and all other columns are copied to the output.
}
