% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text into paragraphs}
\usage{
wrap_text(..., lineLen = 72, indent = 0, exdent = 0)
}
\description{
This program concatenates input strings and collapses consecutive
blanks, and output a single string with inserted newlines.
}
